+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 
+ * Make Caffe work on CPU-only machines ([issue #219](https://github.com/bytedeco/javacpp-presets/issues/219))
  * Fix loading issue with `opencv_face` ([issue bytedeco/javacv#470](https://github.com/bytedeco/javacv/issues/470)
  * Fix presets for CUDA 7.5 on the `linux-ppc64le` platform
- * Upgrade presets for TensorFlow 0.9.0
+ * Upgrade presets for Caffe, TensorFlow 0.9.0
  * Set default options in `tensorflow/cppbuild.sh` to prevent console reads during build
  * Add `Tensor.createStringArray()` method to access `DT_STRING` data ([issue #249](https://github.com/bytedeco/javacpp-presets/issues/249))
  * Fix Javadoc links for externally referenced classes